[Feat] add reload button component
